backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
pydantic==2.11.1
pydantic-settings==2.8.1
python-multipart==0.0.20
openai==1.75.0
pymupdf==1.25.5
python-docx==1.1.2
httpx==0.28.1
pytest==8.3.5
pytest-asyncio==0.26.0
sounddevice==0.5.1
numpy==2.2.4
faster-whisper==1.1.1
fpdf2==2.8.3
